Correct stdc++ language and size issues for MSVC
This patch makes consistent the type for ssize_t as used on Windows.
We will favor describing it as a ptrdiff_t, which is standard C.
Patches to tclap and nghttp2 are needed for this exercise.

Linux is a 64ILP architecture, so there's no difference between int
and [s]size_t and pointer widths. Windows is a 64P architecture, where
only long long maps to the width of the size_t and C99 standard ptrdiff_t
types.

This leads to a number of places where the Envoy API has presented
an explicit int or int32_t which is smaller than the ptrdiff_t on
Windows. In some places we must cast to or from the API defined width.
Several other misassumptions on the width of int are also addressed.

Windows does not support alternately-named __VA_ARGS__ which is not a
standard C++ language feature in our baseline c++ 17 expectation. Other
minor adjustments reflect other quirks of the MSVC compilier (some more
correct and some simply buggy behavior).

Windows MSVC needs various guards on unrecognized #pragmas, and will not
support any preprocessor operations within the arguments to a macro;
this is not stdc or stdc++.

This patch largely ignores changes which are required for windows, and
do not impact the linux/os-x compilation path. This patch is also
missing the IoHandle abstraction of 'fd' arguments, storage and return
codes, win32 error handling and win32 specific #include's handling.

+++ source/extensions/filters/network/redis_proxy/command_splitter_impl.cc
@@ -253,7 +253,7 @@ SplitRequestPtr MGETRequest::create(Router& router, Common::Redis::RespValuePtr&
   single_mget.type(Common::Redis::RespType::Array);
   single_mget.asArray().swap(values);
 
-  for (uint64_t i = 1; i < incoming_request->asArray().size(); i++) {
+  for (unsigned int i = 1; i < incoming_request->asArray().size(); i++) {
     request_ptr->pending_requests_.emplace_back(*request_ptr, i - 1);
     PendingRequest& pending_request = request_ptr->pending_requests_.back();
 
@@ -382,8 +382,8 @@ SplitRequestPtr MSETRequest::create(Router& router, Common::Redis::RespValuePtr&
   single_mset.type(Common::Redis::RespType::Array);
   single_mset.asArray().swap(values);
 
-  uint64_t fragment_index = 0;
-  for (uint64_t i = 1; i < incoming_request->asArray().size(); i += 2) {
+  unsigned fragment_index = 0;
+  for (unsigned i = 1; i < incoming_request->asArray().size(); i += 2) {
     request_ptr->pending_requests_.emplace_back(*request_ptr, fragment_index++);
     PendingRequest& pending_request = request_ptr->pending_requests_.back();
 
@@ -478,7 +478,7 @@ SplitRequestPtr SplitKeysSumResultRequest::create(Router& router,
   single_fragment.type(Common::Redis::RespType::Array);
   single_fragment.asArray().swap(values);
 
-  for (uint64_t i = 1; i < incoming_request->asArray().size(); i++) {
+  for (unsigned i = 1; i < incoming_request->asArray().size(); i++) {
     request_ptr->pending_requests_.emplace_back(*request_ptr, i - 1);
     PendingRequest& pending_request = request_ptr->pending_requests_.back();
